crear nuevo shopping

+++ User/src/main/java/com/proyect/User/repository/ShoppingRepository.java
@@ -4,8 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.proyect.User.model.Shopping;
-import com.proyect.User.response.ShoppingResponse;
 
 @Repository
-public interface ShoppingRepository extends JpaRepository<Shopping, Integer>{
+public interface ShoppingRepository extends JpaRepository<Shopping, Long> {
 }
